Rename shopping-list-card to todo-list-card
