--- gallery/src/pages/lovelace/shopping-list-card.markdown
@@ -1,3 +1,0 @@
----
-title: Shopping List Card
----
