+++ src/translations/en.json
@@ -4471,7 +4471,7 @@
           "entities": {
             "never_triggered": "Never triggered"
           },
-          "shopping-list": {
+          "todo-list": {
             "lists": "To-do Lists",
             "checked_items": "Checked items",
             "clear_items": "Clear checked items",
@@ -5045,9 +5045,9 @@
               "graph_type": "Graph type",
               "description": "The Sensor card gives you a quick overview of your sensors state with an optional graph to visualize change over time."
             },
-            "shopping-list": {
-              "name": "Shopping list",
-              "description": "The Shopping list card allows you to add, edit, check-off, and clear items from your shopping list.",
+            "todo-list": {
+              "name": "Todo list",
+              "description": "The to-do list card allows you to add, edit, check-off, and clear items from your to-do list.",
               "integration_not_loaded": "This card requires the `todo` integration to be set up."
             },
             "thermostat": {
